+++ CHANGELOG.md
@@ -10,6 +10,7 @@ description: |-
 #### v0.9.3 (unreleased)
 - Add secure_key option to cdn resource
 - Add availability_zones and vrrp_port_id fields to vkcs_db_cluster resource
+- Add ability go get admin password of compute instance with windows
 - Support multi-AZ PostgreSQL clusters
 - Fix filtering in the vkcs_cdn_shielding_pop data source
 - Mark api_lb_vip field of vkcs_kubernetes_cluster resource as read-only